+++ src/main/resources/ui/assets/js/f9512c76.5069c08d.js
@@ -1,1 +1,0 @@
-"use strict";(self.webpackChunkvalidator_frontend=self.webpackChunkvalidator_frontend||[]).push([[870],{5745:function(e){e.exports=JSON.parse('{"name":"docusaurus-plugin-content-pages","id":"default"}')}}]);