+++ src/test/java/de/kosit/validationtool/config/ConfigurationBuilderTest.java
@@ -20,15 +20,14 @@ import static de.kosit.validationtool.config.ConfigurationBuilder.report;
 import static de.kosit.validationtool.config.ConfigurationBuilder.schematron;
 import static de.kosit.validationtool.config.TestConfigurationFactory.createSimpleConfiguration;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.Date;
 
-import org.hamcrest.Matchers;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 
 import de.kosit.validationtool.impl.Helper;
 
@@ -41,58 +40,69 @@ public class ConfigurationBuilderTest {
 
     public static final LocalDate EPOCH = LocalDate.of(1970, 1, 1);
 
-    @Rule
-    public ExpectedException exceptions = ExpectedException.none();
-
-    @Test
+    @Test(expected = IllegalStateException.class)
     public void testNoConfiguration() {
-        this.exceptions.expect(IllegalStateException.class);
         new ConfigurationBuilder().build(Helper.getTestProcessor());
     }
 
     @Test
     public void testNoFallback() {
-        this.exceptions.expect(IllegalStateException.class);
-        this.exceptions.expectMessage(Matchers.containsString("fallback"));
-        final ConfigurationBuilder builder = createSimpleConfiguration();
-        builder.with((FallbackBuilder) null);
-        builder.build(Helper.getTestProcessor());
+        try {
+            final ConfigurationBuilder builder = createSimpleConfiguration();
+            builder.with((FallbackBuilder) null);
+            builder.build(Helper.getTestProcessor());
+            fail();
+        } catch (final IllegalStateException ex) {
+            assertTrue(ex.getMessage().contains("fallback"));
+        }
     }
 
     @Test
     public void testNoSchema() {
-        this.exceptions.expect(IllegalStateException.class);
-        this.exceptions.expectMessage(Matchers.containsString("schema"));
-        final ConfigurationBuilder builder = createSimpleConfiguration();
-        builder.getScenarios().get(0).validate((SchemaBuilder) null);
-        builder.build(Helper.getTestProcessor());
+        try {
+            final ConfigurationBuilder builder = createSimpleConfiguration();
+            builder.getScenarios().get(0).validate((SchemaBuilder) null);
+            builder.build(Helper.getTestProcessor());
+            fail();
+        } catch (final IllegalStateException ex) {
+            assertTrue(ex.getMessage().contains("schema"));
+        }
     }
 
     @Test
     public void testInvalidSchematron() {
-        this.exceptions.expect(IllegalStateException.class);
-        this.exceptions.expectMessage(Matchers.containsString("schematron"));
-        final ConfigurationBuilder builder = createSimpleConfiguration();
-        builder.getScenarios().get(0).validate(schematron("invalid").source(URI.create("DoesNotExist")));
-        builder.build(Helper.getTestProcessor());
+        try {
+            final ConfigurationBuilder builder = createSimpleConfiguration();
+            builder.getScenarios().get(0).validate(schematron("invalid").source(URI.create("DoesNotExist")));
+            builder.build(Helper.getTestProcessor());
+            fail();
+        } catch (final IllegalStateException ex) {
+            assertTrue(ex.getMessage().contains("schematron"));
+        }
     }
 
     @Test
     public void testInsufficientSchematron() {
-        this.exceptions.expect(IllegalStateException.class);
-        this.exceptions.expectMessage(Matchers.containsString("schematron"));
-        final ConfigurationBuilder builder = createSimpleConfiguration();
-        builder.getScenarios().get(0).validate(schematron("invalid"));
-        builder.build(Helper.getTestProcessor());
+        try {
+            final ConfigurationBuilder builder = createSimpleConfiguration();
+            builder.getScenarios().get(0).validate(schematron("invalid"));
+            builder.build(Helper.getTestProcessor());
+            fail();
+        } catch (final IllegalStateException ex) {
+            assertTrue(ex.getMessage().contains("schematron"));
+        }
     }
 
     @Test
     public void testNoReport() {
-        this.exceptions.expect(IllegalStateException.class);
-        this.exceptions.expectMessage(Matchers.containsString("report"));
-        final ConfigurationBuilder builder = createSimpleConfiguration();
-        builder.getScenarios().get(0).with(report("invalid"));
-        builder.build(Helper.getTestProcessor());
+        try {
+            final ConfigurationBuilder builder = createSimpleConfiguration();
+            builder.getScenarios().get(0).with(report("invalid"));
+            builder.build(Helper.getTestProcessor());
+            fail();
+        } catch (final IllegalStateException ex) {
+            assertTrue(ex.getMessage().contains("report"));
+        }
     }
 
     @Test